Move authz.SetProviders call to cmd/ package
Found that we call SetProviders in an internal package, if someone decides to call NewStore from another binary we'll unset authz providers which can disable certain auth checks, so that felt a bit risky.

Test plan:

Just moved a bit of code around.
