+++ cmd/syntactic-code-intel-worker/shared/BUILD.bazel
@@ -11,7 +11,11 @@ go_library(
     importpath = "github.com/sourcegraph/sourcegraph/cmd/syntactic-code-intel-worker/shared",
     visibility = ["//visibility:public"],
     deps = [
+        "//internal/authz",
         "//internal/codeintel/syntactic_indexing/jobstore",
+        "//internal/conf",
+        "//internal/conf/conftypes",
+        "//internal/database/connections/live",
         "//internal/debugserver",
         "//internal/encryption/keyring",
         "//internal/env",
